requirements.txt
click==8.1.3
grpcio==1.43.0
grpcio-tools==1.43.0
grpcio-reflection==1.43.0
mypy-protobuf==3.0.0
protobuf==3.18.1
protobuf3==0.2.1
six==1.16.0
types-protobuf==3.20.4.4
kubernetes==24.2.0
kubernetes-validate==1.24.0